+++ src/data/store/key/map/fixed/impls/mod.rs
@@ -1,4 +1,4 @@
-// devela/src/data/store/key/map/impls/mod.rs
+// devela/src/data/store/key/map/fixed/impls/mod.rs
 
 mod r#const; // __map_impl_const!
 mod runtime; // __map_impl_runtime!